move render toast util to ui library
