+++ packages/ui/src/utils/index.ts
@@ -2,3 +2,4 @@
 // SPDX-License-Identifier: GPL-3.0-only
 
 export * from './helpers';
+export * from './toast';
